// src/app/domain/model/project-domain.model.ts
import { ImageAssetDomainType } from '../type/asset/image-asset-domain.type';
import { StatusEntityDomainType } from '../type/entity/status-entity-domain.type';
import { ProjectKindDomainType } from '../type/kind/project-kind-domain.type';
import { TechnologyKindDomainType } from '../type/kind/technology-kind-domain.type';
import { ProjectNameDomainType } from '../type/name/project-name-domain.type';

export namespace ProjectDomain {
  interface RepositoryNoneModel {
    type: 'none';
  }

  interface RepositoryGitHubModel {
    type: 'github';
    url: string;
  }

  type RepositoryModel = RepositoryNoneModel | RepositoryGitHubModel;

  interface ApplicationNoneModel {
    type: 'none';
  }

  interface ApplicationGitHubPageModel {
    type: 'githubPage';
    url: string;
  }

  type ApplicationModel = ApplicationNoneModel | ApplicationGitHubPageModel;

  export interface Data {
    banner: ImageAssetDomainType;
    name: ProjectNameDomainType;
    creationDate: Date;
    status: StatusEntityDomainType;
    description: string;
    repository: RepositoryModel;
    application: ApplicationModel;
    gallery: ImageAssetDomainType[];
  }

  export interface Metadata {
    kind: ProjectKindDomainType;
    technologyIds: TechnologyKindDomainType[];
    order: number;
  }
}